% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cwna_data}
\alias{cwna_data}
\title{Data generated from a constant velocity (or continuous white noise acceleration, CWNA) model for 20 time steps.}
\format{
20 observations of 4 variables.
}
\usage{
cwna_data
}
\description{
Data generated from a constant velocity (or continuous white noise acceleration, CWNA) model for 20 time steps.
}
\keyword{datasets}
