% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{matrix2tidy}
\alias{matrix2tidy}
\title{Convert IS, SMC or EnK output stored as a matrix to tidy format.}
\usage{
matrix2tidy(output, parameter, target = 1, log_weights = NULL)
}
\arguments{
\item{output}{Matrix output (one point per row) from an IS algorithm, or one target from a SMC or EnK algorithm.}

\item{parameter}{The name to assign the parameter in the tidy output.}

\item{target}{(optional) The target index to use in the tidy output (default 1).}

\item{log_weights}{(optional) The log_weights to use in the tidy output (default all equal).}
}
\value{
The output in tidy format.
}
\description{
Convert IS, SMC or EnK output stored as a matrix to tidy format.
}
