% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sir_cwna_model}
\alias{sir_cwna_model}
\title{The output of a bootstrap particle filter on the \code{cwna_data}. The output consists of 100 particles over 20 time steps.}
\format{
4000 observations of 13 variables.
}
\usage{
sir_cwna_model
}
\description{
The output of a bootstrap particle filter on the \code{cwna_data}. The output consists of 100 particles over 20 time steps.
}
\keyword{datasets}
