% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plots.R
\name{scatter_plot}
\alias{scatter_plot}
\title{A histogram of a single variable from a single target.}
\usage{
scatter_plot(
  output,
  x_parameter,
  y_parameter,
  target = NULL,
  pre_weighting = FALSE,
  xlimits = NULL,
  ylimits = NULL,
  default_title = FALSE,
  max_size = 1,
  alpha = 0.1
)
}
\arguments{
\item{output}{Output from the SMC or EnK algorithm.}

\item{x_parameter}{The parameter indexed by the x-axis.}

\item{y_parameter}{The parameter indexed by the y-axis.}

\item{target}{(optional) The index of the target we wish to plot. (default to final target)}

\item{pre_weighting}{(optional) If TRUE, will ignore particle weights in the histogram. If FALSE, will use the particle weights. (defaults to FALSE)}

\item{xlimits}{(optional) Input of the form c(start,end), which specifies the ends of the x-axis.}

\item{ylimits}{(optional) Input of the form c(start,end), which specifies the ends of the y-axis.}

\item{default_title}{(optional) If TRUE, will provide a default title for the figure. If FALSE, no title is used. (defaults to FALSE)}

\item{max_size}{(optional) The maximum size of the points in the plot. (default=1)}

\item{alpha}{(optional) The transparency of the points in the plot. (default=0.1)}
}
\value{
A scatter plot in a ggplot figure.
}
\description{
A histogram of a single variable from a single target.
}
